Increase stack limit for produce
The produce which is processed ores and "fresh fruits and
vegetables" are tracked in the 10 kg units. With the default
stacking limit of 99 items per stack this translates to 1
ton per maximal stack. On the other side a typical ore comes
in a (roughly) 1 ton block and also have stack limit of 99
items per stack which translates into roughly 100 tons per
maximal stack.

This severely cripples the microfactories when recipes with
large outputs are used. For example the water ice
purification (probably the most used as the vast majority
of the ore is water ice) generates 80 units of water per
minute and per block plus 3 other products. With room for
8 stacks this causes the water purificator to become jammed
with clean water in 3-4 minutes, rendering the "automation"
in the microfactories (they can even pull sources from
outputs of anonther node if they don't have them local)
pretty pointless as you can't leave the factory unattended
for too long (even if you carefully balance the whole thing
by adding as many other nodes as needed to consume the
output of the one liquid purificator that processes the
ice.

Challenge is one thing but stupid limitations is NOT a
good challenges. If the game makes the player asking
questions like "I can stuff 100 ton of this into one slot,
why I can't put 100 ton of that in one slot too???", then
something is badly wrong with it.

To balance this increase the stack limit of all produce to
9999 items per stack so a maximal stack holds roughly 100
ton of material similar to ores. As an additional benefit of
this change the microfactories can now be left unattended
for a very long time (a group of 32 liquid purificators can
accept and reprocess the whole hull-load of water ice while
the ship is out filling its hull with another batch) without
getting stuck on full output area or even losing product to
overflow. It doesn't make the factory completely unattended
as you still have to come and collect and sell the product
from time to time but at least you don't have to do thst
every 5 minutes.