Make the station form big and stable
The various station form tabs have varying sizes. The most
glaring problem with that is that whenever you click on a
tab, the tabs themselves jump around annoyingly.

The second problem with the tab sizes is that the forms of
most tabs are too small to fit all the tabs nicely so to
access the "Post Office" or "Hangar And Ship" you must scroll
to these first and then when you go out of them, the tabs
jump back to show the "Equipment market" and you must scroll
again when you want the "Hangar"/"Post Office" ...

Additional size problem is that some of the tabs are way too
small for the content they want to contain, especially the
"Post Office" tab where the "Destination Address" overlaps
with "Due Time" (or whatever that field is named, I can't
read the name at all, only the end of the field gives me a
clue that the field is supposed to contain the package
delivery due time.

The third problem with the sizes is that I want much bigger
markets, more stuff (especially in the Hangar) and a bigger
cargo space for the most expensive hull (and a smaller one
in the most basic form of hull).

This change fixes the "tabs jumping around" and "not all
tabs visible" problems while also adding more space space
for the other stuff I want by making the forms so big they
fill the entire screen (almost). However the form still fits
comfortably the default screen size of Minetest.