Tell the post office to get real with their rewards
Let's admit it, folks, nobody is going to spend a hour
rushing madly and sometimes blindly through the
asteroid-ridden space out here to deliver a parcel for
measly 800 credits when they can spend the same hour digging
said asteroids and selling the resulting cubes to the space
station for three times the profit. And that is not counting
in the microfactories. So tell the post office to increase
their rewards to give their parcels a chance to get moving.
It is still going to be much cheaper for them than to try
to use the Vortex Tunnel Technology.