Put the size indication to one single place
The various forms in the tabbed format *must* hare the size
because otherwise the form is going to behave badly and give
headaches to the player trying to use it. Remove the
temptation to reintroduce the "variable form sizes in tabs"
by moving the size specifier to one single place in the
code and especially by making sure no subroutines calls or
case handling code contains "their own size specifier".