Standardize money and buyout spot
These two items will go through major overhaul in the future
and therefore it will help a lot if there is just one piece
of code that generates these formspecs. Plus there is the
added bonus that the forms look better when the money field
is not jumping around as you switch tabs.