Fix superfluous "info" icons next to hold space
Whenever the hold space is shown in an area where there is
more form space next to the cargo hold inventory, extra
"info" icons are created for the nonexistent inventory spots
in the column just right next to the form. The problem turns
out to be a "off-by-one" bug which was previously hidden by
the fact that in the original version of these forms there
was never "form space next to the right edge of the cargo
hold inventory" and thus these extra info icons got clipped
off.