Fix order of station form display commands
The original code first displayed the form and then
proceeded to setup the metadata in the "ship link" that the
form handling code uses to perform its job.

Somehow this managed to function (most likely because the
function which produces the formspec does not actually query
this metadata store) but it is a serious coding bug making
the code fragile and suspectible to failure if a future
Minetest version decides to change the "showform action"
behavior.

Fix the problem by making sure everything is set correctly
before the form display action is invoked.