Fix office form not being updated properly
The first and most obvious way to see the bug is that when
you look at the clock in the post office, you will see it
does not update. Switch to another tab then switch back and
the clock changes. The clock actually shows the time when
you entered the post office.

A second, subtler but still quite obvious way to see the bug
is to first open a game, interact with the station and go to
the post office. You should see just one package sitting on
the shelves at this point, if not, close the game, reopen it
and go back to the post office. Now wait for a while. A
second package appears on the shelves but there is no
description next to it. Now switch to another tab, switch
back and voila, the description is here.

A third and pretty treacherous way you can be bitten by the
bug occurs when all of the shelves in the post office are
full. To get to that point, open the game and spend a hour
or so mining an asteroid or something. Then return to the
station and (without closing and reopening the game!) go to
the post office. You should see 4 packages sitting on the 4
shelves of the post office. If there is less than 4, exit
the base and spend some more time doing something else. Now
leave the post office form with the 4 packages open for a
long while (hour or teo). After you return to the computer,
take note of the information about the 4 packages. Then
pick up one of the packages. As you do that the information
next to the 3 packages that remained suddenly changes.
After you query the information of the one you picked up
and compare with what the post office said, you will likely
discover that what the package actually says about itself
is not what the post office said about it. Now imagine
grabbing a package with what you thought is an "urgency
class 1" and "destination space station 3", arriving at
the space station 3 and after failing to deliver the
package, checking on it and suddenly realizing that it is
supposed to go to station 4 and is urgency class 4 and
because station 4 is 10 km away and you have only 5
minutes of the delivery term left, you won't make it in
time to get anything more than 10 cr from the delivery
work.

The reason of all these three symptoms is that the formspec
is not updated when either the time or the packages on the
shelves change. This bug gave quite a fight when fixing it
but it should be dead now. The fix involves a periodically
invoked procedure which updates the form when the form is
displaying the post office tab. The hard part was to detect
when the user closed the form. It turned out there is a
provision for that but the station form code failed to
initialize it so it took a significant amount of search to
find it.