Fix crash on overkiller hull downgrade
When changing from overkiller hull to a lesser one, the game
crashes. It turns out this bug was uncovered by the change
that added 12 more cargo slots to that hull and is caused by
the code assuming that the stack position in the inventory
list will hold "nil" when there is nothing at that slot
(newer minetest builds put an empty stack there instead).
Fix this by adding a check that the stack really contains
something before using it to remove stuff from the
inventory.