Remove "free_space" and all code handling it
Since the space in the hull is now handled by the minetest
inventory handling, there is no longer a need to track the
free space in the hull manually, so the field used for that
can now go. That includes the "free space" stat in the hull
registrations as it is now no longer used for anything.