Remove ship "volume" and all code handling it
As volume is now handled automatically by the minetest
inventory framework, there is no need to track the used
volume in the ship stats.

So stop accepting any volume related arguments in the
"apply_cargo" procedure and messing around with the (now
removed) volume stat. Also update all callers to not
bother with passing anything volume related to it anymore.

This change left "calculate_carried_volume" with no use
and no hope to get any use in the future so remove that
too.

After the two changes above the "volume" item in the ship
stats has now no use whatsoever so remove it too.