Make stack limits consistent with volume
This commit now definitely leaves behind all traditions of
Minetest and is the first stage of volume handling
simplification. The main idea is that the "maximal stack
slot" is declared to be 100 cubic meters (or, to be exact,
slightly more to account for any packaging materials but
not much more so let's say that it is 100 m3) and let the
Minetest stack handling code emulate the volume handling.
Since raw ores come in 1m*1m*1m cubes, all ores will have
their stack limit increased to 100 (instead of 99). The
processed ores have their volume stated as 10 dm3 so their
maximal stack limit also needs to be increased from 9999
to 10000.

The next thing to change here is the volumes of the hulls.
This information formerly indicated the cargo space. Now
it indicates the amount of space the hull occupies in the
inventory and as such it was set to 100 for all hulls and
the escape pod.

And last, probably slightly irrelevant but the bullets had
their maximal amount increased from 999 to 10000. Namely I
am not entirely convicted that a bullet for a gun would
have 100 litres. 10 litres also looks pretty excessive but
more realistic. Or, I might increase the price tenfold and
declare the "bullets" to be actually "packs of 100 bullets",
that would reduce the volume of one bullet to reasonable
and realistic 100 millilitres.