Refactor the equipment area
The equipment area in the inventory got labelled and moved
to the left. The "legend" (telling which color was for what
type of equipment) was removed. This new design is far more
compact and far less confusing. As a result this created a
large area on the right of the form, ideal to put the status
info there.

The empty space to the left of the "Radar" box was left
there intentionally and is reserved for the battery slot.
Since there are no batteries implemented yet, there is no
battery slot yet.

Additionally, the "shield" slot was expanded to a massive
2x5 box. The way the shields work will be completely
revamped and brought to sync with the Elite/Oolite games.

The "Hull" equipment box was moved out of the way to a
completely different area below the shields. I am planning
to completely refactor the way hulls are replaced in the
game. However that is another quite complex change that
will take time to develop so for now this "Hull" box is
preserved in this "little corner" so the various
hulls in the game can still be put to use.