Fix annoying inventory delay on startup
When the game is started in the singleplayer mode or a player
connects in the multiplayer mode, in the first 10 seconds
when he tries to access the inventory, a strange
minecraft-like form is shown instead of the correct inventory
and the inventory reverts to normal after a while. If the
player tries to interact with the base during this "broken
inventory form" time, he can't.

The culprit turns out the last line of the "on player
connect" handler that for some mysterious reason tells the
engine to wait for 10 seconds before running the, you
guessed it, inventory reconfiguration part.

It turns out the reason for these delays (two were found, see
the attached patch) was that some parts of the asynchronous
event handling code were overzealously attaching the player
to his ship, causing the player to get attached to multiple
ships after the GUI is loaded and active. The result is
scrambling some GUI state data inside the object, leading to
crashes when the "reattach" happens while the player is
interacting with the space station.