Merge status tab into hull tab
Taking advantage of the new and massive inventory form, move
all of the status information into the hull tab, at the space
prepared.

So add a "status line" between the equipment and the cargo
areas. The first three positions on this status line contain
the cargo hold volume allocation indicators. The fourth
position is occupied by the ship weight and the next two by
the engine strength information. The last position is
reserved for financial information (cash on hand).

To the left of the radar (and above the "status line") there
is a column with information about the power flow within the
ship and next to it is information about the shield energy
flow. These two colums have a number of inactive indicators
(in fact, the power flow column is entirely inactive) because
these two columns are designed for the planned changes in the
gameplay.

The large space left to the right (below the massive 2x5
shields box) is left there intentionally. It will contain a
3x3 box for the "quirium area" which will be shown when the
ship is equipped with the "Quirium Generator And Reactor" in
the battery slot. These two items (the battery slot and the
quirium stuff) are not yet implemented so just an empty
space is at the places where they will once be.

And last but not least, the "Abandon ship" button along with
the code that was handling it, which would completely break
the world if pressed is deleted. It won't be missed.