Fix map becoming "stale" after player moves
Go to the inventory screen and switch to the map. Exit the
inventory screen. Start your engines and move somewhere else.
Go back to the inventory. The map should show up but it will
still show the old location of the player. Only after
switching the tabs the map updates itself.

The problem is that the formspec becoms outdated if the form
is supposed to show the map and the player moves. To fix the
problem, the various "set_player_inventory_formspec" calls
around the code are removed and replaced with a timer
procedure that invokes the set_player_inventory_formspec
periodically 10 times per second no matter what is happening
to the player.

However care needs to be taken to not do this periodic form
updating when the info form is shown as that would make it
impossible to show the info form at all. Therefore a new
state variable is introduced which tracks whether there is
an info form open so the periodic form updater can make sure
the info form is left intact.