Fix space station form messing up inventory tab
After the user interacts with the space staton, the inventory
form forgets which tab it had selected. If the player left
the space station form while viewing one of the three
markets, the inventory form will show one of the three
inventory form tabs. If the station interaction was
terminated while one of the non-market tabs was being shown,
the inventory form will show an empty page and the player is
required to select one of the three tabs to make it work.

The problem turns out to be that the two forms use the same
ship state variable to remember which tab is selected and
the space station interaction code does not bother restoring
the value of the ship state variable to what the inventory
form had there. Adding this missing functionality fixes the
problem.