Refactor localization and unit descriptions
The formats are now splitted between two tables. The first
one specifies the "normal properties" (or "basic properties")
and the second one specifies the "special properties", which
acts like "overlay" i.e. only what is different from "normal
properties" is mentioned in these items. The code that
displays the properties now assembles the "row info" from
the apropriate tables ("normal properties" table only for
"basic properties", both tables for "special properties") and
calls a function to do the actual row construction.
This removes most of the code duplication that was present in
the original code.

Also a special case of "date" was replaced by the ability
to specify a formatting procedure rather than a format.
This is done by giving a format_proc instead of
format_numeric (or no format at all) which points to the
function to use for the formatting. The function is supposed
to take a value and returns its string representation.

This facility replaces the special value "date" which
triggers formatting of the value as a date. The two "date"
fields were converted to use the new formatting procedure
facility instead.