Fix inability to get info in post office
The info screen fails to come up if the player clicks an info
icon in his cargo hold inventory. The problem is the same as
with the inventory screen and the info icons there: a
periodic form updater squashes the old form without paying
attention to the fact that it might be the info form.