More readable repair price calculation
After this it becomes clear what kind of rip-off the repair
shop here actually is. It turns out repair costs about 6
times more than to simply buy another brand new item, use
the old one until it breaks and then place the new at its
place.

The only problem with the "more economic" variant is that
if the engines break when you are flying into an asteroid
and need to stop or change direction pretty quickly, you
might not have the time to swap a brand new engine in
before the asteroid turns you into stardust.