Remove middle cube in anchor indicator
It appears the code that shows the indicator of the anchor
protected area was copied from some protection mod. The
middle cube in that indicator was used to "surround the
protector" so the user would know to which one the shown area
actually belongs. In this game the protector can be anywhere
within the block and there is only one per player so the
middle cube serves no useful purpose, therefore remove it.