Identify the sides of the display
Instead of stupid "sides" comment lumping the 4 sides
together (and making it pretty much impossible to figure out
which is which, mark each side with the name of the direction
of the vector coming from the center of the cube towards the
side. This makes editing of the thing much easier.